+++ ChangeLog.md
@@ -12,6 +12,7 @@
  * raise error on illegal string escapes
  * fix parser bug preventing multi-module files
  * allow empty module export lists
+ * fix segfault in pure morloc @load intrinsic
 
 0.83.2 [2026-05-07]
 -------------------